First pass at leveling consolidation
