+++ MatterControl.csproj
@@ -75,6 +75,8 @@
     <Compile Include="AboutPage\CheckForUpdatesPage.cs" />
     <Compile Include="ApplicationView\LogoSpinner.cs" />
     <Compile Include="ConfigurationPage\CalibrationSettings\CalibrationControls.cs" />
+    <Compile Include="ConfigurationPage\PrintLeveling\LevelWizard3x3Mesh.cs" />
+    <Compile Include="ConfigurationPage\PrintLeveling\MeshLevlingFunctions.cs" />
     <Compile Include="ConfigurationPage\PrintLeveling\ProbeCalibrationWizard.cs" />
     <Compile Include="CustomWidgets\ConfigurePrinterWidget.cs" />
     <Compile Include="AboutPage\HTMLParser\HtmlWidget.cs" />
@@ -225,7 +227,6 @@
     <Compile Include="ConfigurationPage\LanguageSelector.cs" />
     <Compile Include="ConfigurationPage\PrintLeveling\InstructionsPage.cs" />
     <Compile Include="ConfigurationPage\PrintLeveling\LevelingStrings.cs" />
-    <Compile Include="ConfigurationPage\PrintLeveling\LevelWizardMesh.cs" />
     <Compile Include="ConfigurationPage\PrintLeveling\LevelWizard7PointRadial.cs" />
     <Compile Include="ConfigurationPage\PrintLeveling\LevelWizard13PointRadial.cs" />
     <Compile Include="ConfigurationPage\PrintLeveling\LevelWizard3Point.cs" />
